+++ dipole_inverse/cython_lib/pop_matrix_lib.pyx
@@ -6,8 +6,10 @@ cdef extern from "pop_matrix_C_lib.h":
 
     void populate_matrix_C(double * G,
                            double * QDM_domain, double scan_height,
-                           double * cuboids, int N_cuboids, int Npart,
-                           int Ny, int Nx, double QDM_spacing,
+                           double * cuboids,
+                           unsigned long long N_cuboids, unsigned long long Npart,
+                           unsigned long long Ny, unsigned long long Nx, 
+                           double QDM_spacing,
                            double QDM_deltax, double QDM_deltay,
                            int Origin, int verbose
                            )
@@ -17,8 +19,11 @@ cdef extern from "pop_matrix_C_lib.h":
 def populate_matrix_cython(double [:, :] G,
                            double [:] QDM_domain,
                            double scan_height,
-                           double [:] cuboids, int N_cuboids,
-                           int Npart, int Ny, int Nx, double QDM_spacing,
+                           double [:] cuboids, 
+                           unsigned long long N_cuboids,
+                           unsigned long long Npart, 
+                           unsigned long long Ny, unsigned long long Nx,
+                           double QDM_spacing,
                            double QDM_deltax, double QDM_deltay,
                            int Origin, int verbose):
